add side menu as child of button
