+++ app/src/components/Layout.tsx
@@ -3,17 +3,10 @@ import TopBar from './TopBar'
 import React from 'react'
 
 export const Layout = () => {
-  const [sideDrawerOpen, setSideDrawerOpen] = React.useState<boolean>(false);
-
-  const toggleSideDrawer = () => {
-    setSideDrawerOpen(true);
-  };
-
   return (
     <>
-    <TopBar toggleSideDrawer={toggleSideDrawer}/>
+    <TopBar/>
     <Outlet />
-    <p>sideDrawerOpen: {sideDrawerOpen.toString()}</p>
     </>
   )
 }
