+++ app/src/components/TopBar.tsx
@@ -5,11 +5,9 @@ import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import SideDrawerButton from './SideDrawerButton';
 
-interface Props {
-  toggleSideDrawer: () => void;
-}
 
-const TopBar: React.FC<Props> = ({ toggleSideDrawer }) => {
+
+const TopBar: React.FC = () => {
   return (
     <Box sx={{ flexGrow: 1 }}>
       <AppBar position="static">
@@ -17,7 +15,8 @@ const TopBar: React.FC<Props> = ({ toggleSideDrawer }) => {
           <Typography variant="h6" sx={{ flexGrow: 1 }}>
             HapTrack
           </Typography>
-          <SideDrawerButton toggleSideDrawer={toggleSideDrawer} />
+          
+          <SideDrawerButton/>
         </Toolbar>
       </AppBar>
     </Box>
